Load time from nano seconds
Load time is now a double instead of long. I am internally using
System.nanoTime() instead of System.currentTimeMillis(). The time
difference is still in milliseconds.

+++ src/com/flipkart/fk_android_flipperf/models/PerfDataSet.java
@@ -145,7 +145,7 @@ public class PerfDataSet {
 		public Long startTime;
 
 		@SerializedName("loadTime")
-		public Long loadTime;
+		public Double loadTime;
 
 		@SerializedName("endTime")
 		public Long endTime;
